Rename /src/css to /src/styles to prevent confusion
