+++ src/styles/Icon.scss
