+++ src/styles/StatusIcon.scss
