+++ src/styles/TextEllipsis.scss
