+++ src/styles/_animations.scss
